created package "yoeo" in vision_modules
