+++ bitbots_vision/bitbots_vision/vision_modules/yoeo/__init__.py
@@ -1,0 +1,4 @@
+from .object_manager import YOEOObjectManager
+from .vision_components import IVisionComponent, YOEOComponent, CameraCapCheckComponent, YOEOBallDetectionComponent, \
+    YOEOGoalpostDetectionComponent, YOEOFieldBoundaryDetectionComponent, YOEOLineDetectionComponent, \
+    YOEOFieldDetectionComponent, YOEOObstacleDetectionComponent, DebugImageComponent
